refactor(LayoutFlow): extract graph logic into custom hooks and components
